requirements.txt
openai==1.35.7
python-dotenv==1.0.1
fastapi==0.111.0
aiogram==3.8.0
supabase==2.5.1
fluent.runtime==0.4.0
fluent.syntax==0.19.0
structlog==24.2.0
google-generativeai==0.8.3
pillow==11.0.0
